Fixed blank text box
App would crash if there text box was blank. Fixed it.

Changed the length of the toast messages to be shorter.

Changed the distance from double to just float.

+++ app/src/main/java/com/example/getgpslocation/GPSTracker.java
@@ -149,7 +149,7 @@ public class GPSTracker extends Service implements LocationListener{
 	}
 
 	//Finds distance between point A and point B with latitude and longitude
-	public double getDistanceTo(double latA, double lngA, double latB, double lngB) {
+	public float getDistanceTo(double latA, double lngA, double latB, double lngB) {
 		Location locationA = new Location("Point A");
 
 		locationA.setLatitude(latA);
